moved named multimesh to from the input component to the multimesh component
