+++ components/input/src/wmtk/components/input/CMakeLists.txt
@@ -10,13 +10,6 @@ set(SRC_FILES
     mesh_with_tag_from_image.hpp
     input.cpp
 	input.hpp
-    NamedMultiMesh.hpp
-    NamedMultiMesh.cpp
-    utils/get_attribute.hpp
-    utils/get_attribute.cpp
-
-    MeshCollection.hpp
-    MeshCollection.cpp
 )
 
 include(stb)
@@ -25,5 +18,6 @@ include(stb)
 target_compile_features(wmtk_${COMPONENT_NAME} PUBLIC cxx_std_20)
 
 target_sources(wmtk_${COMPONENT_NAME} PRIVATE ${SRC_FILES})
-target_link_libraries(wmtk_${COMPONENT_NAME} PRIVATE stb::image)
+target_link_libraries(wmtk_${COMPONENT_NAME} PRIVATE stb::image nlohmann_json)
+target_link_libraries(wmtk_${COMPONENT_NAME} PRIVATE stb::image nlohmann_json PUBLIC wmtk::multimesh)
