+++ components/input/src/wmtk/components/input/input.hpp
@@ -4,7 +4,7 @@
 
 #include <wmtk/Mesh.hpp>
 
-#include "NamedMultiMesh.hpp"
+#include <wmtk/components/multimesh/NamedMultiMesh.hpp>
 namespace wmtk::components::input {
 class InputOptions;
 
@@ -25,6 +25,6 @@ std::shared_ptr<Mesh> input(
     const bool ignore_z_if_zero = false,
     const std::vector<std::string>& tetrahedron_attributes = {});
 
-NamedMultiMesh input(const InputOptions& options);
+multimesh::NamedMultiMesh input(const InputOptions& options);
 
 } // namespace wmtk::components::input
