+++ components/input/tests/CMakeLists.txt
@@ -1,2 +1,2 @@
 
-add_component_test(wmtk::${COMPONENT_NAME} input.cpp named_multimesh.cpp)
+add_component_test(wmtk::${COMPONENT_NAME} input.cpp )
