+++ components/multimesh/CMakeLists.txt
@@ -1,3 +1,3 @@
 set(COMPONENT_NAME multimesh)
 add_subdirectory("src/wmtk/components/${COMPONENT_NAME}")
-#add_subdirectory("tests")
+add_subdirectory("tests")
