+++ components/multimesh/src/wmtk/components/multimesh/CMakeLists.txt
@@ -20,6 +20,15 @@ set(SRC_FILES
 
     multimesh.hpp
     multimesh.cpp
+
+    NamedMultiMesh.hpp
+    NamedMultiMesh.cpp
+    utils/get_attribute.hpp
+    utils/get_attribute.cpp
+    internal/split_path.hpp
+
+    MeshCollection.hpp
+    MeshCollection.cpp
 )
 
 include(stb)
@@ -28,3 +37,4 @@ include(stb)
 target_compile_features(wmtk_${COMPONENT_NAME} PUBLIC cxx_std_20)
 
 target_sources(wmtk_${COMPONENT_NAME} PRIVATE ${SRC_FILES})
+target_link_libraries(wmtk_${COMPONENT_NAME} PRIVATE stb::image nlohmann_json)
