+++ components/multimesh/src/wmtk/components/multimesh/MeshCollection.hpp
@@ -2,14 +2,14 @@
 #include <map>
 #include "NamedMultiMesh.hpp"
 
-namespace wmtk::components::input {
+namespace wmtk::components::multimesh {
 
     class InputOptions;
 class MeshCollection
 {
 public:
     NamedMultiMesh& add_mesh(NamedMultiMesh o);
-    NamedMultiMesh& add_mesh(const InputOptions& opts);
+    //NamedMultiMesh& add_mesh(const InputOptions& opts);
 
     const NamedMultiMesh& get_named_multimesh(const std::string_view& path) const;
     const Mesh& get_mesh(const std::string_view& path) const;
@@ -22,4 +22,4 @@ public:
 private:
     std::map<std::string_view, NamedMultiMesh> m_meshes;
 };
-} // namespace wmtk::components::input
+} // namespace wmtk::components::multimesh
