+++ components/multimesh/tests/CMakeLists.txt
@@ -1,0 +1,4 @@
+
+add_component_test(wmtk::${COMPONENT_NAME} named_multimesh.cpp)
+
+target_link_libraries(${WMTK_COMPONENT_TEST_TARGET} PRIVATE wmtk::input)
